+++ flumine/markets/blotter.py
@@ -72,7 +72,10 @@ class Blotter:
     @property
     def live_orders(self) -> bool:
         for order in self._orders.values():
-            if order.status == OrderStatus.EXECUTABLE:
+            if (
+                order.status == OrderStatus.EXECUTABLE
+                or order.trade.trade_complete is False
+            ):
                 return True
         return False
 
@@ -89,7 +92,9 @@ class Blotter:
 
     def selection_exposure(self, strategy, lookup: tuple) -> float:
         """Returns strategy/selection exposure,
-        max value is 0
+        can be positive or negative.
+            positive = profit on selection
+            negative = exposure on selection
         """
         mb, ml = [], []  # (price, size)
         for order in self: